fix: marketo util code refactor to accomodate marketo static list
