+++ __tests__/data/marketo_router_output.json
@@ -240,7 +240,7 @@
   {
     "batched": false,
     "statusCode": 400,
-    "error": "Request Failed for Marketo, Lookup field 'userId' not found (Aborted).[Marketo Transformer]: During lead look up using email",
+    "error": "Request Failed for marketo, Lookup field 'userId' not found (Aborted).[Marketo Transformer]: During lead look up using email",
     "statTags": {
       "destType": "MARKETO",
       "stage": "transform",
