+++ __tests__/data/marketo_static_list_router_output.json
@@ -143,7 +143,7 @@
         "statusCode": 400,
         "error": "track call is not supported.",
         "statTags": {
-            "destType": "MARKETO",
+            "destType": "MARKETO STATIC LIST",
             "meta": "badEvent",
             "scope": "transformation",
             "stage": "transform"
